+++ services/frontend/www-app/src/components/SingleModeListItem.vue
@@ -7,7 +7,7 @@
 <script lang="ts">
 import Route from 'src/models/Route';
 import { defineComponent, PropType } from 'vue';
-import { TravelmuxTrip } from 'src/services/TravelmuxClient';
+import Trip from 'src/models/Trip';
 
 export default defineComponent({
   name: 'SingleModeListItem',
@@ -19,7 +19,7 @@ export default defineComponent({
   },
   props: {
     trip: {
-      type: Object as PropType<TravelmuxTrip>,
+      type: Object as PropType<Trip>,
       required: true,
     },
     // SingleModalListItem actually doesn't use this, but MultiModalListItem needs it, so
