+++ services/frontend/www-app/src/components/SingleModeSteps.vue
@@ -39,13 +39,13 @@ import {
   valhallaTypeToIcon,
 } from 'src/services/ValhallaClient';
 import { getBaseMap } from './BaseMap.vue';
-import { TravelmuxTrip } from 'src/services/TravelmuxClient';
+import Trip from 'src/models/Trip';
 
 export default defineComponent({
   name: 'SingleModeSteps',
   props: {
     trip: {
-      type: Object as PropType<TravelmuxTrip>,
+      type: Object as PropType<Trip>,
       required: true,
     },
   },
